Refactor: extract noItems component for list
