+++ src/components/organisms/recipe/detail/stepsList/index.jsx
@@ -1,23 +1,12 @@
 import React, { useState } from 'react';
 import PropTypes from 'prop-types';
 
-import { makeStyles } from '@material-ui/core';
-
-import Typography from '../../../../atoms/typography.jsx';
+import NoItems from '../../../list/noItems.jsx';
 
 import Section from './components/section.jsx';
 import Item from './components/item.jsx';
 
 const Component = ({ className, steps }) => {
-  const useStyles = makeStyles(theme => ({
-    message: {
-      color: theme.palette.primary.dark,
-      textAlign: 'center',
-      paddingBottom: '20px',
-    },
-  }));
-  const classes = useStyles();
-
   let rowCounter = -1;
 
   const [checked, setChecked] = useState([]);
@@ -62,10 +51,10 @@ const Component = ({ className, steps }) => {
       />
     ))
   ) : (
-    <Typography variant="body1" className={classes.message}>
-      No steps required
-    </Typography>
-  );
+      <NoItems>
+        No steps required
+      </NoItems>
+    );
 };
 
 Component.defaultProps = {
